pyproject.toml
[project]
name = "thermal-printer"
version = "0.1.0"
description = "Thermal printer web interface"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "python-escpos",
    "pillow",
    "python-multipart",
    "pillow-heif",
    "pyusb>=1.3.1",
    "pycups>=2.0.4",
]
